% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecan-package.R
\docType{package}
\name{pecan-package}
\alias{pecan}
\alias{pecan-package}
\title{pecan: Partial Eta-Squared for Crossed and Nested Linear Mixed Models}
\description{
Computes partial eta-squared effect sizes for fixed effects in linear mixed models fitted with lme4. Supports crossed and nested random effects structures with any number of grouping factors. Random slope variances are translated to the outcome scale using a variance decomposition approach, correctly accounting for predictor scaling and interaction terms. Both general and operative effect sizes are provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcohen0901/pecan}
  \item Report bugs at \url{https://github.com/bcohen0901/pecan/issues}
}

}
\author{
\strong{Maintainer}: Brandon Cohen \email{brandon.cohen-1@colorado.edu} (\href{https://orcid.org/0009-0004-1802-3435}{ORCID})

}
\keyword{internal}
