% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pecanr-package.R
\docType{package}
\name{pecanr-package}
\alias{pecanr}
\alias{pecanr-package}
\title{pecanr: Partial Eta-Squared for Crossed and Nested Linear Mixed Models}
\description{
Computes partial eta-squared effect sizes for fixed effects in linear mixed models fitted with the 'lme4' package. Supports crossed and nested random effects structures with any number of grouping factors. Random slope variances are translated to the outcome scale using a variance decomposition approach, correctly accounting for predictor scaling and interaction terms. Both general and operative effect sizes are provided. Methods are based on Correll, Mellinger, McClelland, and Judd (2020) \doi{10.1016/j.tics.2019.12.009}, Correll, Mellinger, and Pedersen (2022) \doi{10.3758/s13428-021-01687-2}, and Rights and Sterba (2019) \doi{10.1037/met0000184}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/bcohen0901/pecanr}
  \item Report bugs at \url{https://github.com/bcohen0901/pecanr/issues}
}

}
\author{
\strong{Maintainer}: Brandon Cohen \email{brandon.cohen-1@colorado.edu} (\href{https://orcid.org/0009-0004-1802-3435}{ORCID})

Authors:
\itemize{
  \item Joshua Correll \email{joshua.correll@colorado.edu} (\href{https://orcid.org/0000-0003-1142-459X}{ORCID}) [thesis advisor]
}

}
\keyword{internal}
